+++ packages/avilatek_core/lib/avilatek_core.dart
@@ -1,6 +1,0 @@
-/// A Very Good Project created by Very Good CLI.
-library avilatek_core;
-
-export 'common/core.dart';
-export 'http/http.dart';
-export 'interfaces/interfaces.dart';