+++ packages/avilatek_core/lib/common/core.dart
@@ -1,2 +1,5 @@
+/// Core utilities commonly used in Avila Tek projects.
+library core;
+
 export 'assertions.dart';
 export 'error_radar.dart';