+++ packages/avilatek_core/lib/http/http.dart
@@ -1,3 +1,6 @@
+/// HTTP client and utilities commonly used in Avila Tek projects.
+library http;
+
 export 'headers_injector.dart';
 export 'persistent_headers_injector.dart';
 export 'session_client.dart';