+++ packages/avilatek_core/lib/interfaces/interfaces.dart
@@ -1,1 +1,5 @@
+/// Interfaces for most common use cases in Avila Tek projects, and
+/// for other πackages of avilatek_core.
+library interfaces;
+
 export 'device_storage.dart';