+++ packages/avilatek_core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.0.0
 Major breaking changes were introduced in this version.
-* refactor: (BREAKING) new package folder structure. The package is now split up into multiple modules. As of now, the package contains the following modules: common, http and interfaces.
+* refactor: (BREAKING) new package structure. avilatek_core is now split up into multiple package modules. As of now, the package contains the following modules: common, http and interfaces.
 * refactor: (BREAKING) renamed `IErrorRadarDelegate` to `ErrorRadarDelegate`, `MultiErrorRadarRepository` to `MultiErrorRadar` and removed `IErrorRadarRepository` entirely along with ErrorRadarRepository. 
 * refactor: (BREAKING) renamed `HttpHeadersInjector` class and its methods. The class was renamed to `HeadersInjector`  
 * feat: added `PersistentHeadersInjector` class to handle HTTP headers persistence.